refactor: move each model services hooks into dedicated file
