+++ src/services/model/index.ts
@@ -1,0 +1,27 @@
+import useCreateGithubModel from "./useCreateGithubModel";
+import useCreateLocalModel from "./useCreateLocalModel";
+import useDeployModelInstance from "./useDeployModelInstance";
+import useModel from "./useModel";
+import useModelDefinition from "./useModelDefinition";
+import useModelDefinitions from "./useModelDefinitions";
+import useModelInstance from "./useModelInstance";
+import useModelInstances from "./useModelInstances";
+import useModels from "./useModels";
+import useModelsWithInstances from "./useModelsWithInstances";
+import useModelWithInstances from "./useModelWithInstances";
+import useUpdateModel from "./useUpdateModel";
+
+export {
+  useCreateGithubModel,
+  useCreateLocalModel,
+  useDeployModelInstance,
+  useModel,
+  useModelDefinition,
+  useModelDefinitions,
+  useModelInstance,
+  useModelInstances,
+  useModels,
+  useModelsWithInstances,
+  useModelWithInstances,
+  useUpdateModel,
+};
