+++ src/services/model/useModelInstances.ts
@@ -1,0 +1,22 @@
+import { useQuery } from "react-query";
+import { listModelInstancesQuery } from "@/lib/instill";
+import { Nullable } from "@/types/general";
+
+const useModelInstances = (modelName: Nullable<string>) => {
+  return useQuery(
+    ["models", modelName, "modelInstances"],
+    async () => {
+      if (!modelName) {
+        return Promise.reject(new Error("Model name not provided"));
+      }
+
+      const modelInstances = await listModelInstancesQuery(modelName);
+      return Promise.resolve(modelInstances);
+    },
+    {
+      enabled: modelName ? true : false,
+    }
+  );
+};
+
+export default useModelInstances;
