+++ src/services/model/useModels.ts
@@ -1,0 +1,11 @@
+import { useQuery } from "react-query";
+import { listModelsQuery } from "@/lib/instill";
+
+const useModels = () => {
+  return useQuery(["models"], async () => {
+    const models = await listModelsQuery();
+    return Promise.resolve(models);
+  });
+};
+
+export default useModels;
